base43 -> util
